change state according to dynamics requirement
